Introducing long, short, and byte types! (and long literals)
* Adapted type system according to Java language spec, now stuff like 'c' + 3 is allowed
* Added Long, Short, and Byte classes to std lib
* Removed special Random class, which is not needed anymore, and updated all benchmarks

+++ montecarlo/src/main/java/ch/trick17/rolezapps/montecarlo/MonteCarloAppJava.java
@@ -53,7 +53,7 @@ public class MonteCarloAppJava extends MonteCarloApp {
     
     private final Returns returns;
     
-    private final int[] seeds;
+    private final long[] seeds;
     private final List<Double> results = new ArrayList<>();
 
     public MonteCarloAppJava(String ratesFile, int steps, int runs, int numTasks, long $task) {
@@ -62,7 +62,7 @@ public class MonteCarloAppJava extends MonteCarloApp {
         
         returns = new Returns(RatePath.readRatesFile(ratesFile));
         
-        seeds = new int[runs];
+        seeds = new long[runs];
         for(int i = 0; i < runs; i++)
             seeds[i] = i * 11;
     }
